+++ test/airport/standardRoute/RouteSegmentCollection.spec.js
@@ -2,9 +2,9 @@
 import ava from 'ava';
 import _isEqual from 'lodash/isEqual';
 
-import RouteSegmentCollection from '../../../src/assets/scripts/airport/StandardRoute/RouteSegmentCollection';
-import RouteSegmentModel from '../../../src/assets/scripts/airport/StandardRoute/RouteSegmentModel';
-import FixCollection from '../../../src/assets/scripts/airport/Fix/FixCollection';
+import RouteSegmentCollection from '../../../src/assets/scripts/client/airport/StandardRoute/RouteSegmentCollection';
+import RouteSegmentModel from '../../../src/assets/scripts/client/airport/StandardRoute/RouteSegmentModel';
+import FixCollection from '../../../src/assets/scripts/client/airport/Fix/FixCollection';
 
 import { airportPositionFixtureKSFO } from '../../fixtures/airportFixtures';
 import { FIX_LIST_MOCK } from '../fix/_mocks/fixMocks';