+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 3.3.0 (January 20, 2017)
+---
 ### Major
-
+- Restructures `src` files into `client` and `server` folders. [#220](https://github.com/n8rzz/atc/issues/220)